#38
- удален useOkoPush так как подобные искусственные ограничения не помогают улучшить код
- мелкие правки стиля кода в utils/browser/analytics
