+++ src/utils/browser/analytics/data-layer.ts
@@ -3,8 +3,10 @@
  * @param eventData Данные события для отправки.
  */
 export function dataLayerPush(eventData: Record<string, any>): void {
-  const win: any = window;
+  const target: any = window;
 
-  // копируем объект так как window.dataLayer.push может менять аргумент в процессе работы
-  typeof win.dataLayer?.push === 'function' && win.dataLayer.push({ ...eventData });
+  if (typeof target.dataLayer?.push === 'function') {
+    // копируем объект так как window.dataLayer.push может менять аргумент в процессе работы
+    target.dataLayer.push({ ...eventData });
+  }
 }
